Changes for slicing utils and LabelledArray
